Extract calendar event actions into impl
